Update UI chart
