+++ superset/assets/src/explore/components/ExploreChartHeader.jsx
@@ -25,8 +25,6 @@ import ExploreActionButtons from './ExploreActionButtons';
 import RowCountLabel from './RowCountLabel';
 import EditableTitle from '../../components/EditableTitle';
 import AlteredSliceTag from '../../components/AlteredSliceTag';
-import FaveStar from '../../components/FaveStar';
-import TooltipWrapper from '../../components/TooltipWrapper';
 import Timer from '../../components/Timer';
 import CachedLabel from '../../components/CachedLabel';
 
@@ -38,12 +36,9 @@ const CHART_STATUS_MAP = {
 
 const propTypes = {
   actions: PropTypes.object.isRequired,
-  addHistory: PropTypes.func,
-  can_overwrite: PropTypes.bool.isRequired,
   can_download: PropTypes.bool.isRequired,
   isStarred: PropTypes.bool.isRequired,
   slice: PropTypes.object,
-  table_name: PropTypes.string,
   form_data: PropTypes.object,
   timeout: PropTypes.number,
   chart: chartPropShape,
@@ -55,37 +50,6 @@ class ExploreChartHeader extends React.PureComponent {
       this.props.timeout, this.props.chart.id);
   }
 
-  updateChartTitleOrSaveSlice(newTitle) {
-    const isNewSlice = !this.props.slice;
-    const params = {
-      slice_name: newTitle,
-      action: isNewSlice ? 'saveas' : 'overwrite',
-    };
-    this.props.actions.saveSlice(this.props.form_data, params)
-      .then((json) => {
-        const { data } = json;
-        if (isNewSlice) {
-          this.props.actions.updateChartId(data.slice.slice_id, 0);
-          this.props.actions.createNewSlice(
-            data.can_add, data.can_download, data.can_overwrite,
-            data.slice, data.form_data);
-          this.props.addHistory({ isReplace: true, title: `[chart] ${data.slice.slice_name}` });
-        } else {
-          this.props.actions.updateChartTitle(newTitle);
-        }
-      });
-  }
-
-  renderChartTitle() {
-    let title;
-    if (this.props.slice) {
-      title = this.props.slice.slice_name;
-    } else {
-      title = t('%s - untitled', this.props.table_name);
-    }
-    return title;
-  }
-
   render() {
     const formData = this.props.form_data;
     const {
@@ -94,40 +58,12 @@ class ExploreChartHeader extends React.PureComponent {
       chartUpdateStartTime,
       latestQueryFormData,
       queryResponse } = this.props.chart;
-      const chartFinished = ['failed', 'rendered', 'success'].includes(this.props.chart.chartStatus);
+    const chartFinished = ['failed', 'rendered', 'success'].includes(this.props.chart.chartStatus);
     return (
       <div
         id="slice-header"
         className="clearfix panel-title-large"
       >
-        <EditableTitle
-          title={this.renderChartTitle()}
-          canEdit={!this.props.slice || this.props.can_overwrite}
-          onSaveTitle={this.updateChartTitleOrSaveSlice.bind(this)}
-        />
-
-        {this.props.slice &&
-        <span>
-          <FaveStar
-            itemId={this.props.slice.slice_id}
-            fetchFaveStar={this.props.actions.fetchFaveStar}
-            saveFaveStar={this.props.actions.saveFaveStar}
-            isStarred={this.props.isStarred}
-          />
-
-          <TooltipWrapper
-            label="edit-desc"
-            tooltip={t('Edit chart properties')}
-          >
-            <a
-              className="edit-desc-icon"
-              href={`/chart/edit/${this.props.slice.slice_id}`}
-            >
-              <i className="fa fa-edit" />
-            </a>
-          </TooltipWrapper>
-        </span>
-        }
         {this.props.chart.sliceFormData &&
           <AlteredSliceTag
             origFormData={this.props.chart.sliceFormData}
