+++ superset/assets/stylesheets/less/overfit/variables.less
@@ -20,3 +20,7 @@
 @row-table-color-hover: #e5f3f5;
 @row-table-color-active: #d2f3f5;
 @border-dropdown: rgba(25, 138, 206, 0.1);
+@text-grey-light: #99a2a8;
+@btn-primary-active: #136ca1;
+@chart-title: #b8c3ca;
+@btn-run-query: #43b734;
