+++ superset/templates/superset/add_slice.html
@@ -19,19 +19,35 @@
 {% extends "superset/basic.html" %}
 
 {% block title %}
-  Add new slice
+Add new slice
 {% endblock %}
 
 {% block body %}
-  <div
-    id="js-add-slice-container"
-    data-bootstrap="{{ bootstrap_data }}"
-  ></div>
+<div class="container-fluid">
+  <div class="row custom-layout">
+    <div class="menu-left">
+      <ul class="nav navbar-nav">
+        {% if WARNING_MSG %}
+        <li class="alert alert-danger">
+          {{ WARNING_MSG | safe }}
+        </li>
+        {% endif %}
+        {% include 'appbuilder/menu.html' %}
+      </ul>
+    </div>
+    <div class="boby-content">
+      <div id="js-add-slice-container" data-bootstrap="{{ bootstrap_data }}">
+        <img src="/static/assets/images/loading.gif"
+          style="width: 50px; position: absolute; top: 50%; left: 50%; transform: translate(-50%,-50%)">
+      </div>
+    </div>
+  </div>
+</div>
 {% endblock %}
 
 {% block tail_js %}
-  {{ super() }}
-  {% with filename="addSlice" %}
-    {% include "superset/partials/_script_tag.html" %}
-  {% endwith %}
+{{ super() }}
+{% with filename="addSlice" %}
+{% include "superset/partials/_script_tag.html" %}
+{% endwith %}
 {% endblock %}
